backend/requirements.txt
ultralytics>=8.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pillow>=10.0.0
python-multipart>=0.0.6
opencv-python>=4.8.0
websockets>=12.0
yt-dlp>=2024.0.0
httpx>=0.25.0
imageio-ffmpeg>=0.4.9
onnxruntime-directml>=1.16.0
